FRONTEND: cleanup MaterialColor usage
the problem is that the initialization order of static members is not defined. So the color array was filled with garbage,
because the video::Color::XXX members (also static) were not yet set. This behaviour could have worked with one linker run
and could have been broken with another one - as it depends on the compiler and the order of execution

the solution is to lazy initialize the material color array by doing a scoped static array in a function