VIDEO: introduced shader cvars and allow to hot reload shaders
if you create a shader cvar with the flag CV_SHADER the app is automatically reloading all shaders
and put #defines into the shader code for every CV_SHADER flagged cvar. This can be used to prevent
branching in the shaders and use the preprocessor to activate/deactivate different code paths

The included example is not complete, cl_debug_ambientocclusion is not fully implemented in the
shader - but opening the console and set cl_debug_ambientocclusion to true and then to false again
should show you the power of this feature.

You can also bind this to a command (there will hopefully soon also be a toggle command that you
can bind to a key)