VIDEO: update call while vao is bound is not allowed anymore
it looks like this can destroy the vao state in some cases with some drivers (sorry for not being more
specific ;) ). Calling unbindBuffer while vao is bound breaks things for the imgui rendering (there both
the vertex and the index buffer are updated). not calling unbindbuffer breaks the voxel editor (but this
is most likely an internal state error, not a driver bug)