VOXEDIT: moved modifier execution into action button
this reduced code duplication and also moved the event handling out of the ui code

also introduced an input context value. This value can be used to determine whether
a given command should get executed. E.g. a command that should only be valid when
the game scene has the focus (like jumping trigger) but a bindings for the same key
should execute a different command in e.g. the ui input context.