ANIMATION: new module for voxel skeletal character animation
the bones are defined in code
the stock system is used to render the active tool
the attrib system is used to resolve the speed value

added new test tool called testanimation

the frontend module got support for voxel characters
and was also introduced into the mapview tool.

there are currently three models and a few items as well
as a glider. More will follow.

animations inspired by cube world