VOXELFORMAT: allow to apply transform when exporting to obj
some formats like vxm apply transformations that can't get mapped to
integer volume boundaries. We don't apply the transformation to the
volume for this format (other as e.g. for magicavoxel vox - were it
does apply to integer boundaries) - but you have the option to apply
the transformation either in voxedit by inspecting the scene graph
or when exporting to obj or ply by specifying the new cvar to do so.