VOXELFORMAT: MESH: reduced memory footprint for mesh voxelization
converted from glm::vec4 to core::RGBA (reduction is 8 bytes per pos entry)

changed testVoxelizeColor by adding a third color
voxelizeTris is now performing a lookup instead of re-using the palette index
for the newly added color. This fixes a problem with chosing the wrong colors
in a few situations.
only sample the center of the triangle
reduced the float multiplications a lot to speed up color handling