VOXELFORMAT: GLTF: prepare to improved material support for gltf and magicavoxel interactions
In order to support the new palette material features that are modelled after magicavoxel (see
1a9c6763fd9e7f87b080da31dbc1e682e59773a6) we have to use some extensions to the gltf spec.

* KHR_materials_emissive_strength
* KHR_materials_ior
* KHR_materials_pbrSpecularGlossiness
* KHR_materials_specular

there are also some new cvars that will control the gltf export extensions - when loading a
model, we will try to load as many extensions as useful automatically. But for saving the
user might have to choose between several options. Some of them can not be used together.

see issue #402