VOXELFORMAT: VOX: missing palette color #549
this improved the situation where a full 256 color palette should get mapped onto a palette that only supports 255 colors
because one of the colors is representing air (e.g. sandbox vxm or magicavoxel vox) - we are now not just looking for
a suitable replacement to get the good result, but we also try to first search for an unused slot in the palette that we
are then using for the slot that represents air in the target format/palette.

to work around this, make sure to use a properly formed palette file. E.g. for magicavoxel you should use the first slot
for a fully transparent pixel to indicate an empty slot - for sandbox vxm you should use the 256st color to be fully
transparent. If you assemble your palette that way you won't get any reordering or color reduction